Worked around a Blizzard bug where GetSpellPowerCost() doesn't always set the cost correctly.
Sometimes, with Levitate for exampe, the table that GetSpellPowerCost() returns has the .cost member set to 0.  This is inaccurate, so to work around it, I calculate the cost myself using the costPercent member and UnitPower.

This is still slightly inaccurate, however, since the costPercent variable appears to be wrong.  Using Levitate as an example again, costPercent is set to 1, when in reality, the spell appears to cost 0.9% of base mana.

Whatever, it doesn't need to be exact.